Add Supabase integration and new allocation features
- Introduced .env.example for Supabase configuration.
- Updated package.json and package-lock.json to include Supabase dependencies.
- Created API routes for managing allocation snapshots.
- Added allocation and history pages to display and manage allocations.
- Enhanced state management in the store to support new allocation features.
- Implemented UI components for allocation management.
- Added necessary database migration for allocation snapshots.
