+++ .env.example
@@ -1,0 +1,2 @@
+SUPABASE_URL=
+SUPABASE_SERVICE_ROLE_KEY=
