+++ app/allocation/page.tsx
@@ -1,0 +1,5 @@
+import AllocationPage from "@/components/tranche/allocation-page";
+
+export default function AllocationRoute() {
+  return <AllocationPage />;
+}
