+++ app/layout.tsx
@@ -1,5 +1,5 @@
 import type { Metadata } from "next";
-import { IBM_Plex_Sans, JetBrains_Mono } from "next/font/google";
+import { Bebas_Neue, IBM_Plex_Sans, JetBrains_Mono } from "next/font/google";
 
 import { Toaster } from "@/components/ui/sonner";
 import "./globals.css";
@@ -16,6 +16,12 @@ const jetBrainsMono = JetBrains_Mono({
   variable: "--font-mono",
 });
 
+const bebasNeue = Bebas_Neue({
+  subsets: ["latin"],
+  weight: "400",
+  variable: "--font-logo",
+});
+
 export const metadata: Metadata = {
   title: "Tranche",
   description: "Stock allocation tool.",
@@ -27,7 +33,7 @@ export default function RootLayout({
   children: React.ReactNode;
 }>) {
   return (
-    <html lang="en" className={`${ibmPlexSans.variable} ${jetBrainsMono.variable} dark`}>
+    <html lang="en" className={`${ibmPlexSans.variable} ${jetBrainsMono.variable} ${bebasNeue.variable} dark`}>
       <body className="[font-family:var(--font-ui)] antialiased">
         {children}
         <Toaster richColors position="top-right" />
