+++ supabase/.temp/pooler-url
@@ -1,0 +1,1 @@
+postgresql://postgres.dycjkahnnjfgsulpkfpu@aws-1-us-east-1.pooler.supabase.com:5432/postgres
